requirements.txt
Django==2.2.6
pytz==2019.3
whitenoise==4.1.4
gunicorn==20.0.0
psycopg2==2.7.5
djangorestframework==3.10.3
dj-config-url==0.1.1
dj-database-url==0.5.0
django-crispy-forms==1.7.2
django-cors-headers==3.2.0